docs(contributing): add Contributing Guide
Took the challenge guide from the main repo and reworked it for the Curriculum repo. Moved the
CONTRIBUTING.md file to the root directory. Added an image directory and sample image to more
clearly show how to open a PR and point it to the correct branch, like in the guide in the
main repo.

ISSUES CLOSED: #61